Fixed error where user could not delete profile. It was because we tried to delete something that didn't exist

+++ src/middlewares/deleteImage.js
@@ -4,7 +4,7 @@ const staticFolder = path.join(process.cwd(), "static");
 
 module.exports = function (userId) {
     const userFiles = path.join(staticFolder, userId);
-
+    if (!fs.existsSync(userFiles)) return true;
     const dirContent = fs.readdirSync(userFiles);
     dirContent.forEach((file) => fs.unlinkSync(path.join(userFiles, file)));
     fs.rmdirSync(userFiles);
